Adding a .gitignore Solving all compilation (not linking) warnings level 3 generated by updater and its dependencies (VS2015 Upd3) Upgrading tinyxml solution from VS 2010 to VS 2015

+++ external/AnyOption/anyoption.cpp
@@ -956,7 +956,7 @@ AnyOption::readFile( const char* fname )
                 return NULL;
         }
         is.seekg (0, ios::end);
-        length = is.tellg();
+        length = static_cast<int>(is.tellg());
         is.seekg (0, ios::beg);
         buffer = (char*) malloc(length*sizeof(char));
         is.read (buffer,length);
